requirements.txt
openai>=1.0.0
python-dotenv
fastapi
uvicorn

beautifulsoup4
requests
faiss-cpu
tiktoken
chromadb

langgraph
langchain-core



